backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pandas==2.2.3
numpy==2.1.3
sqlalchemy==2.0.36
yfinance==0.2.50
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.3
aiofiles==24.1.0
httpx==0.28.1
matplotlib==3.9.3
scipy==1.14.1
alpha-vantage==2.3.1